src/pyproject.toml
[project]
name = "execution-v1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
    "aiosmtplib>=3.0.0",
    "aiosqlite>=0.19.0",
    "cryptography>=41.0.0",
    "icecream>=2.1.8",
    "jinja2>=3.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyinstaller>=6.16.0",
    "python-dotenv>=1.0.0",
    "pytz>=2023.3",
    "selenium>=4.37.0",
    "selenium-async>=0.1.0",
    "sqlalchemy>=2.0.0",
    "sqlmodel>=0.0.14",
]
